backend/pyproject.toml
[tool.poetry]
name = "Waterwise API"
version = "0.1.0"
description = "API for mobile application 'Waterwise'"
authors = ["dpnam2112@gmail.com"]
license = "MIT"
readme = "README.md"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.4"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
toml = "^0.10.2"
loguru = "^0.7.2"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
redis = "^5.0.4"
supervisor = "^4.2.5"
httpx = "^0.27.2"
tenacity = "^9.0.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
cryptography = "^44.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
